Move getGeneProp to utils file
