+++ app/views/ReportView/components/PharmacoGenomicSummary/columnDefs.ts
@@ -1,12 +1,5 @@
 import ArrayCell from '@/components/DataTable/components/ArrayCellRenderer';
-
-const getGeneProp = (params, property) => {
-  const { data: { variant, variantType } } = params;
-  if (variantType === 'sv') {
-    return variant.gene1[property] || variant.gene2[property] || false;
-  }
-  return variant.gene[property] || false;
-};
+import getGeneProp from '@/utils/getGeneProp';
 
 const sampleColumnDefs = [
   {
